arm64: dts: qcom: Remove unused smb1355
The SMB1355 is not used on FP3, so remove it from dts.

We still need to keep pmi632_ctm pinctrl enabled, as charger and camera
flash LED depend on this. Also for good measure keep the pmi632_sense
configuration to avoid unexpected surprises.

See also the following commits:
* 15d3c8f9991e ("ARM: dts: msm: enable H/W thermal mitigation for PMI632 based platform")
* ca13a0ce9d35 ("ARM: dts: msm: Add PMI632 GPIO configuration for parallel-current sense")

Issue: FP3-A13#71
Change-Id: I841a2090af7b72082c8c9a19b2ba044f944190ba